Reorganized everything.
All the recipe information is in the recipe-collection page and the instruction displays the detail of a recipe depending on which one the user clicks in the previous page.

Problem is since I use the same line to import gif and video in the instruction html, we have to decide between gif or video. Now I used ion-img so only gifs are displayed and the videos in Xinyu's recipe do not show.

The ingredlist might be helpful to create a shopping list, we can think more about that.